[android] Moved analytics interfaces to pack com.mapswithme.analytics
# Conflicts:
#	android/src/com/mapswithme/maps/ExternalLibrariesMediator.java
