+++ android/src/com/mapswithme/maps/ExternalLibrariesMediator.java
@@ -13,10 +13,9 @@ import com.crashlytics.android.ndk.CrashlyticsNdk;
 import com.google.android.gms.ads.identifier.AdvertisingIdClient;
 import com.google.android.gms.common.GooglePlayServicesNotAvailableException;
 import com.google.android.gms.common.GooglePlayServicesRepairableException;
-import com.mapswithme.maps.content.DefaultEventLogger;
-import com.mapswithme.maps.content.EventLogger;
-import com.mapswithme.maps.content.EventLoggerAggregator;
-import com.mapswithme.util.SharedPropertiesUtils;
+import com.mapswithme.maps.analytics.DefaultEventLogger;
+import com.mapswithme.maps.analytics.EventLogger;
+import com.mapswithme.maps.analytics.EventLoggerAggregator;
 import com.mapswithme.util.Utils;
 import com.mapswithme.util.log.Logger;
 import com.mapswithme.util.log.LoggerFactory;
