+++ android/src/com/mapswithme/maps/analytics/FlurryEventLogger.java
@@ -1,4 +1,4 @@
-package com.mapswithme.maps.content;
+package com.mapswithme.maps.analytics;
 
 import android.app.Activity;
 import android.app.Application;
